Move NullLocation to a separate class
